+++ types.d.ts
@@ -47,6 +47,7 @@ export namespace Monti {
     function trackError(error: Error, options?: TrackErrorOptions): void;
 
     function ignoreErrorTracking(error: Error): void;
+    function keepProcessAlive(error: Error): void;
 
     function startEvent(name: string, data?: Record<string, any>, fn?: Function): MontiEvent | false;
     function endEvent(event: MontiEvent | false, data?: Record<string, any>): void;